+++ Block/Form/AdditionalInfo.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Copyright © Alekseon sp. z o.o.
+ * http://www.alekseon.com/
+ */
+namespace Alekseon\WidgetForms\Block\Form;
+
+use Magento\Framework\View\Element\Template;
+
+/**
+ * Class AdditionalInfo
+ * @package Alekseon\WidgetForms\Block\Form
+ */
+class AdditionalInfo extends \Magento\Framework\View\Element\Template
+{
+    /**
+     * @var string
+     */
+    protected $_template = 'Alekseon_WidgetForms::form/additional_info.phtml';
+}